[refactoring] #206 Move checkDuplicateNickname method to Member API
