+++ core/network/src/main/java/com/cmc/curtaincall/core/network/service/member/MemberService.kt
@@ -5,6 +5,7 @@ import com.cmc.curtaincall.core.network.service.member.request.DeleteMemberReque
 import com.cmc.curtaincall.core.network.service.member.request.MemberCreateRequest
 import com.cmc.curtaincall.core.network.service.member.request.UpdateMemberRequest
 import com.cmc.curtaincall.core.network.service.member.response.MemberCreateResponse
+import com.cmc.curtaincall.core.network.service.member.response.MemberDuplicateNickNameResponse
 import com.cmc.curtaincall.core.network.service.member.response.MemberInfoResponse
 import com.cmc.curtaincall.core.network.service.member.response.MemberLostItemsResponse
 import com.cmc.curtaincall.core.network.service.member.response.MemberParticipationsResponse
@@ -21,6 +22,11 @@ import retrofit2.http.Path
 import retrofit2.http.Query
 
 interface MemberService {
+    @GET("members/duplicate/nickname")
+    suspend fun checkDuplicateNickname(
+        @Query("nickname") nickname: String
+    ): MemberDuplicateNickNameResponse
+
     @POST("signup")
     suspend fun createMember(
         @Body memberCreateRequest: MemberCreateRequest
