+++ data/src/main/java/com/cmc/curtaincall/data/source/remote/AuthRemoteSource.kt
@@ -1,6 +1,7 @@
 package com.cmc.curtaincall.data.source.remote
 
 import com.cmc.curtaincall.core.network.service.auth.AuthService
+import com.cmc.curtaincall.core.network.service.auth.request.LoginRequest
 import com.cmc.curtaincall.core.network.service.auth.response.LoginResponse
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
@@ -9,8 +10,8 @@ import javax.inject.Inject
 class AuthRemoteSource @Inject constructor(
     private val authService: AuthService
 ) {
-    fun requestLogin(authorization: String): Flow<LoginResponse> = flow {
-        emit(authService.requestLogin(authorization = "Bearer $authorization"))
+    fun requestLogin(provider: String, token: String): Flow<LoginResponse> = flow {
+        emit(authService.requestLogin(provider, LoginRequest(token)))
     }
 
     fun requestReissue(refreshToken: String): Flow<LoginResponse> = flow {
@@ -19,9 +20,5 @@ class AuthRemoteSource @Inject constructor(
 
     fun requestLogout(accessToken: String): Flow<Boolean> = flow {
         emit(authService.requestLogout("Bearer $accessToken").success)
-    }
-
-    fun checkDuplicateNickname(nickname: String): Flow<Boolean> = flow {
-        emit(authService.checkDuplicateNickname(nickname).result)
     }
 }
