+++ data/src/main/java/com/cmc/curtaincall/data/source/remote/MemberRemoteSource.kt
@@ -14,6 +14,9 @@ import javax.inject.Inject
 class MemberRemoteSource @Inject constructor(
     private val memberService: MemberService
 ) {
+    fun checkDuplicateNickname(nickname: String): Flow<Boolean> = flow {
+        emit(memberService.checkDuplicateNickname(nickname).result)
+    }
     fun createMember(nickname: String): Flow<Int> = flow {
         emit(memberService.createMember(MemberCreateRequest(nickname)).id)
     }
