+++ domain/src/main/java/com/cmc/curtaincall/domain/repository/AuthRepository.kt
@@ -4,11 +4,7 @@ import com.cmc.curtaincall.domain.model.auth.LoginResultModel
 import kotlinx.coroutines.flow.Flow
 
 interface AuthRepository {
-    fun requestLogin(authorization: String): Flow<LoginResultModel>
-
+    fun requestLogin(provider: String, token: String): Flow<LoginResultModel>
     fun requestReissue(refreshToken: String): Flow<LoginResultModel>
-
     fun requestLogout(accessToken: String): Flow<Boolean>
-
-    fun checkDuplicateNickname(nickname: String): Flow<Boolean>
 }
