+++ domain/src/main/java/com/cmc/curtaincall/domain/repository/MemberRepository.kt
@@ -9,8 +9,8 @@ import com.cmc.curtaincall.domain.model.member.MyRecruitmentModel
 import kotlinx.coroutines.flow.Flow
 
 interface MemberRepository {
+    fun checkDuplicateNickname(nickname: String): Flow<Boolean>
     fun createMember(nickname: String): Flow<Int>
-
     fun requestMemberInfo(memberId: Int): Flow<MemberInfoModel>
 
     fun updateMember(
